Clean up setup and teardown logic
