+++ swift-spaces-real-time-app/MuxSpacesUIKit/View/ParticipantsViewModel+Space.swift
@@ -2,46 +2,94 @@
 //  ParticipantsViewModel+Space.swift
 //
 
+import Combine
 import Foundation
 import MuxSpaces
 
 extension ParticipantsViewModel {
 
-    // MARK: - Convenience constructor
+    // MARK: - Setup Observers on Space State Updates
 
-    static func make(
-        with token: String,
-        audioCaptureOptions: AudioCaptureOptions?,
-        videoCaptureOptions: CameraCaptureOptions?
-    ) -> ParticipantsViewModel? {
+    func setupEventHandlers() {
+        /// Setup observers that will update your views state
+        /// based on events that are produced by the space
+        ///
+        space
+            .events
+            .joinSuccesses
+            .sink { [weak self] _ in
+                guard let self = self else {
+                    return
+                }
 
-        // Check that the token is not empty
-        // before proceeding
-        guard !token.isEmpty else {
-            return nil
-        }
+                self.handleJoinSuccess()
+            }
+            .store(in: &cancellables)
 
-        // Initialize a Space with a pre-generated token
-        guard let space = try? Space(
-            token: token
-        ) else {
-            return nil
-        }
-
-        // Initialize a view model that will translate
-        // Space state changes to changes in the app UI
-        let viewModel = ParticipantsViewModel(
-            space: space,
-            audioCaptureOptions: audioCaptureOptions,
-            cameraCaptureOptions: videoCaptureOptions
+        Combine.Publishers.Merge5(
+            /// Participant joined events trigger a new
+            /// cell to be added for each new participant
+            space.events
+                .participantJoined
+                .map(\.participant.connectionID),
+            /// When the SDK subscribes to a new video track,
+            /// the participants video becomes available to display
+            space.events
+                .videoTrackSubscriptions
+                .map(\.participant.connectionID),
+            /// When the SDK unsubscribes from a video track,
+            /// the participants video should be taken down
+            /// this update is handled in ParticipantsViewModel
+            space.events
+                .videoTrackUnsubscriptions
+                .map(\.participant.connectionID),
+            // We only want to know when our own tracks are
+            // published
+            space.events
+                .videoTrackPublications
+                .filter { $0.participant.isLocal }
+                .map(\.participant.connectionID),
+            // We only want to know when our own tracks are
+            // published
+            space.events
+                .videoTrackUnpublications
+                .filter { $0.participant.isLocal }
+                .map(\.participant.connectionID)
         )
+        .sink(receiveValue: { [weak self] (connectionID: Participant.ConnectionID) in
+            guard let self = self else { return }
 
-        return viewModel
+            self.snapshot.upsertParticipant(
+                connectionID
+            )
+        })
+        .store(in: &cancellables)
+
+        /// Each participant leaving will cause the applicable cell
+        /// to be removed.
+        space.events
+            .participantLeft
+            .map(\.participant.connectionID)
+            .sink(receiveValue: { [weak self] (connectionID: Participant.ConnectionID) in
+                guard let self = self else { return }
+
+                self.snapshot.removeParticipant(
+                    connectionID
+                )
+            })
+            .store(in: &cancellables)
+    }
+
+    func tearDownEventHandlers() {
+        /// Tear down observers
+        cancellables.forEach { $0.cancel() }
     }
 
     // MARK: - Join space handling
 
     func joinSpace() {
+        setupEventHandlers()
+
         space.join()
     }
 
@@ -67,7 +115,7 @@ extension ParticipantsViewModel {
         /// down the camera and mic
         space.leave()
 
-        cancellables.forEach { $0.cancel() }
+        tearDownEventHandlers()
     }
 
     // MARK: - Publish and Unpublish Tracks
